Update user data store and menu item network service organization
