+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/composables/onboarding/CheckOnboardingViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.jjerrell.meta.course.sample.littlelemon.data.provideUserDataSource
+import app.jjerrell.meta.course.sample.littlelemon.network.provideUserDataSource
 import kotlinx.coroutines.launch
 
 class CheckOnboardingViewModel : ViewModel() {
