+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/composables/onboarding/OnboardingViewModel.kt
@@ -7,8 +7,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.jjerrell.meta.course.sample.littlelemon.data.model.UserRegistration
-import app.jjerrell.meta.course.sample.littlelemon.data.provideUserDataSource
+import app.jjerrell.meta.course.sample.littlelemon.network.model.UserRegistration
+import app.jjerrell.meta.course.sample.littlelemon.network.provideUserDataSource
 import kotlinx.coroutines.launch
 
 class OnboardingViewModel : ViewModel() {
