--- app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/data/model/MenuItemNetwork.kt
@@ -1,21 +1,0 @@
-package app.jjerrell.meta.course.sample.littlelemon.data.model
-
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MenuItemNetwork(
-    @SerialName("category")
-    val category: String,
-    @SerialName("description")
-    val description: String,
-    @SerialName("id")
-    val id: Int,
-    @SerialName("image")
-    val image: String,
-    @SerialName("price")
-    val price: String,
-    @SerialName("title")
-    val title: String
-)
