+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/network/model/MenuItemNetwork.kt
@@ -1,0 +1,21 @@
+package app.jjerrell.meta.course.sample.littlelemon.network.model
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MenuItemNetwork(
+    @SerialName("id")
+    val id: Int,
+    @SerialName("title")
+    val title: String,
+    @SerialName("description")
+    val description: String,
+    @SerialName("price")
+    val price: String,
+    @SerialName("image")
+    val image: String,
+    @SerialName("category")
+    val category: String
+)
